+++ src/LoadableApp.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import Loadable from "react-loadable";
 const LoadApp = Loadable({
-  loader: () => import("./speciesLink/validator"),
+  loader: () => import("./speciesLink/validator_speciesLink"),
   loading() {
     return <div>Loading...</div>;
   },
